requirements.txt
pywebview>=5.0
httpx>=0.27
fpdf2>=2.7
openpyxl>=3.1
mcp>=1.2
pymupdf4llm>=1.27
pyyaml>=6.0
google-auth>=2.30
google-auth-oauthlib>=1.2
google-api-python-client>=2.140
